delete

+++ lan-life/src/main/java/file/dao/RegularDao.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author lan
  * @date 2019-03-04 02:00
  */
-public interface RegularDao extends JpaRepository<Regular, Integer> {
+public interface RegularDao extends JpaRepository<Regular, Long> {
 
 
 }
